fix:入力が行えない不具合を修正
onBlurプロパティとuseStateを使用し、表示用の値のステートとpropsから渡されるステートを分けて処理
